Moved tests for datacube.utils.generic into separate file
test_utils.py is way too large
